refactor lap_tools
